pyproject.toml
[project]
name = "dscopilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.3.8",
    "duckdb>=1.4.2",
    "fastapi[standard]>=0.124.0",
    "google-genai>=1.53.0",
    "groq>=0.37.1",
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pip>=25.3",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
